Feat: divided release table section from info tab
